docker-practice-app/app/requirements.txt
flask==3.0.3
flask-cors==4.0.1
redis==5.0.4
psycopg2-binary==2.9.9
PyMySQL==1.1.1
cryptography==42.0.8